ai-server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
torch
numpy
pandas
scikit-learn
joblib
pytest
pytest-cov
alembic
openai
python-multipart
sqlalchemy

textblob
httpx
orjson
python-dateutil
fastapi-cors
loguru
vaderSentiment
pytest
pytest-asyncio
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
openai>=1.3.0
textblob==0.17.1
spacy>=3.7.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==23.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
textblob==0.17.1
spacy==3.7.2
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1